// frontend/src/api/generated/shop/schemas/paginatedOrdersListResponseList.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * ShopDjango API
 * API for ShopDjango project
 * OpenAPI spec version: 1.0.0
 */
import type { OrdersListResponse } from './ordersListResponse';

export interface PaginatedOrdersListResponseList {
  count: number;
  /** @nullable */
  next?: string | null;
  /** @nullable */
  previous?: string | null;
  results: OrdersListResponse[];
}